Bug fixes about launch abort and timer overflow
Fixed a bug where if the programmer specifies a status to trigger the
status will be switched at t=0

+++ ROCKET_SEQUENCER/ROCKET_SEQUENCER.h
@@ -12,7 +12,7 @@
 #define ROCKET_SEQUENCER_H
 
 // this is a test comment
-typedef unsigned short (*func_compare_t) (void);
+typedef unsigned long (*func_compare_t) (void);
 typedef void (*function_noarg_t) (void); 		
 typedef void (*function_onearg_t) (unsigned short);
 typedef void (*function_2arg_t) (unsigned short, unsigned short);
@@ -34,7 +34,7 @@ typedef struct { // pointers to functions to call before and up to rocket's each
 		function_onearg_t function_onearg;
 		function_2arg_t function_twoarg;};
 	func_compare_t func_to_compare;
-	union{short cmp2; status_t status;};
+	union{unsigned long cmp2; status_t status;};
 	unsigned char arg1;
 	unsigned char arg2;
 	compare_t sign;
@@ -48,7 +48,7 @@ extern status_t _status;
 extern unsigned long trigger_zero;//record rocket start time (in respect to program's start time)
 extern unsigned short countdown_time;//countdown time (in milliseconds)
 extern int triggerPin;
-extern unsigned short timer; // the rocket sequencer timer in milliseconds (both count down and count up)
+extern unsigned long timer; // the rocket sequencer timer in milliseconds (both count down and count up)
 
 extern void initialize(unsigned char stages, int trigger_pin = -1); //define rocket stages and trigger pin
 extern void setCountdownTime(unsigned short time); // set countdown time(in milliseconds)
@@ -67,14 +67,14 @@ extern void loadSequence(unsigned stage_num,
 			  		      * countdown due to a plane from homeland security fouling
 			  		      * the range)
 			  		      */
-extern void set_condition(func_compare_t cmp1, short cmp2,
+extern void set_condition(func_compare_t cmp1, unsigned long cmp2,
 		compare_t sign_input, unsigned stage_num); // set the event's starting condition
 extern void set_condition(status_t status_to_compare,
 		compare_t sign_input, unsigned stage_num); //function overload for statuses
 extern void start(); // start the sequencer (THE FLIGHT COMPUTER HAS TAKEN OVER THE COUNTDOWN)
 extern status_t getStatus(); // GET the status of the rocket (as defined in enum status_t)
 extern void setStatus(status_t status); // set the status of the rocket
-extern unsigned short getTimer(); // get the time in milliseconds from the rocket timer
+extern unsigned long getTimer(); // get the time in milliseconds from the rocket timer
